+++ hutool-db/src/main/java/cn/hutool/db/SqlConnRunner.java
@@ -20,6 +20,7 @@ import cn.hutool.db.handler.PageResultHandler;
 import cn.hutool.db.handler.RsHandler;
 import cn.hutool.db.sql.Query;
 import cn.hutool.db.sql.SqlExecutor;
+import cn.hutool.db.sql.Wrapper;
 import cn.hutool.db.sql.Condition.LikeType;
 import cn.hutool.log.StaticLog;
 
@@ -550,8 +551,30 @@ public class SqlConnRunner{
 	 * 设置SQL方言
 	 * @param dialect 方言
 	 */
-	public void setDialect(Dialect dialect) {
+	public SqlConnRunner setDialect(Dialect dialect) {
 		this.dialect = dialect;
+		return this;
+	}
+	
+	/**
+	 * 设置包装器，包装器用于对表名、字段名进行符号包装（例如双引号），防止关键字与这些表名或字段冲突
+	 * @param wrapperChar 包装字符，字符会在SQL生成时位于表名和字段名两边，null时表示取消包装
+	 * @return this
+	 * @since 4.0.0
+	 */
+	public SqlConnRunner setWrapper(Character wrapperChar) {
+		return setWrapper(new Wrapper(wrapperChar));
+	}
+	
+	/**
+	 * 设置包装器，包装器用于对表名、字段名进行符号包装（例如双引号），防止关键字与这些表名或字段冲突
+	 * @param wrapper 包装器，null表示取消包装
+	 * @return this
+	 * @since 4.0.0
+	 */
+	public SqlConnRunner setWrapper(Wrapper wrapper) {
+		this.dialect.setWrapper(wrapper);
+		return this;
 	}
 	//---------------------------------------------------------------------------- Getters and Setters end